src/python-storage/requirements.txt
azure-core==1.31.0
azure-identity==1.19.0
azure-storage-blob==12.23.1